Move groupbykey to grouping
[ghstack-poisoned]
